Add support for specifying the Growl alert title
Support using the RECORD SEPARATOR (036) control character
to split the growl message string into multiple parts to
more explicitly control the growl behavior. With this change,
the first part is used as the alert title and the second part
is the message, mostly backwards-compatible with prior
releases:
    "\e]9;Alert Title \036Growl Notification\007"

If no record separator is present, the entire string is used
as the user's message as before.

Records past the first two are currently ignored for future
compatibility.